+++ fake_winreg/fake_reg.py
@@ -1,7 +1,7 @@
 # STDLIB
 import platform
 import time
-from typing import Dict, Optional, Union
+from typing import Dict, List, Optional, Union
 
 
 is_windows = platform.system().lower() == 'windows'
@@ -54,7 +54,7 @@ class FakeRegistryValue(object):
         # the name of the value
         self.value_name: str = ''
         # the value
-        self.value: Union[None, bytes, str, int] = ''
+        self.value: Union[None, bytes, str, List[str], int] = ''
         # the REG_* type of the Value
         self.value_type: int = REG_SZ
         # used in module fake_winreg, Default = KEY_READ
@@ -127,7 +127,7 @@ def get_fake_reg_key(fake_reg_key: FakeRegistryKey, sub_key: str) -> FakeRegistr
 
 
 def set_fake_reg_value(fake_reg_key: FakeRegistryKey, sub_key: str,
-                       value_name: str, value: Union[None, bytes, str, int], value_type: int = REG_SZ,
+                       value_name: str, value: Union[None, bytes, str, List[str], int], value_type: int = REG_SZ,
                        last_modified_ns: Union[int, None] = None) -> FakeRegistryValue:
     """
     sets the value of the fake key - we create here keys on the fly, but beware of the last_modified_ns time !