+++ CHANGES.rst
@@ -7,9 +7,9 @@ Changelog
 
 0.2.1
 -----
-2020-xx-xx : development branch
-    -
-    -
+2020-07-12 : feature release
+    - raise Errors on SetValueEx if type is not appropriate
+    - more documentation
 
 
 0.2.0